+++ src/observer/ActiveTextEditorChangeObserver.ts
@@ -3,6 +3,7 @@ import Observer from './Observer';
 import PluginClassDecorationProvider from 'decorator/PluginClassDecorationProvider';
 import Context from 'common/Context';
 import ObserverInstanceDecorationProvider from 'decorator/ObserverInstanceDecorationProvider';
+import CronClassDecorationProvider from 'decorator/CronClassDecorationProvider';
 
 export default class ActiveTextEditorChangeObserver extends Observer {
   public async execute(textEditor: TextEditor | undefined): Promise<void> {
@@ -11,14 +12,17 @@ export default class ActiveTextEditorChangeObserver extends Observer {
     if (textEditor && textEditor.document.languageId === 'php') {
       const pluginProvider = new PluginClassDecorationProvider(textEditor.document);
       const observerProvider = new ObserverInstanceDecorationProvider(textEditor.document);
+      const cronProvider = new CronClassDecorationProvider(textEditor.document);
 
-      const [pluginDecorations, observerDecorations] = await Promise.all([
+      const [pluginDecorations, observerDecorations, cronDecorations] = await Promise.all([
         pluginProvider.getDecorations(),
         observerProvider.getDecorations(),
+        cronProvider.getDecorations(),
       ]);
 
       textEditor.setDecorations(pluginProvider.getType(), pluginDecorations);
       textEditor.setDecorations(observerProvider.getType(), observerDecorations);
+      textEditor.setDecorations(cronProvider.getType(), cronDecorations);
     }
   }
 }