+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 - Added: Event name autocomplete
 - Added: Hovering CRON job schedules will show a human readable version
+- Added: Cron job indexer and instance class decorations
 - Changed: Implemented batching for the indexer to reduce load
 
 ## [1.5.0] - 2025-04-06